[#252 state:fixed] Completely removed MainWindow.shown_account.
This completes the transition from a shared AccountView to one instance of AccountView per opened Account pane. This shared AccountView thing was the cause of the entry selection glitch in #252. With one account instance per account, the entry selection doesn't have to be restored because it isn't touched when we switch accounts.