[#262 state:fixed] Schedules are now updated by global changes when appropriate.
When a global change becomes the first spawn of a schedule, it replaces the ref txn of the schedule. This is to avoid the confusing situations where the schedule transaction would be different from all its spawns, and that changing a schedule would have no effect on its spawn (a new global change would be required).

Also, deleting the first spawn(s) of a schedule will make its start date move forward.