[#325 state:fixed] Changed the label of the 3 date swap items in the import window.
Instead of just telling which fields they swap, they display the current and if-fixed date formats, thus improving the clarity of the operation. Moreover, the default date format of the system is tried first (unless the parsed format has a native date format) before trying the whole list of date formats.

That was a lot of work for a seemingly simple feature...