[#339] Merged the two pie charts in networth and profit view.
This one pie chart per view now manages the double-pie drawing logic. This lead to a lot of simplification on the Cocoa side, but it also required a lot of changes in the chart printing logic because the old logic of painting the view directly in the print view wouldn't work (it relied on NSView copying, which can't work since the drawing is now made in the core and that copying a view in Cocoa doesn't result in copying a chart in the core).